requirements.txt
langchain-openai>=0.1.8,<0.2.0
python-dotenv>=1.0.1,<2.0.0
langchain>=0.2.1,<0.3.0
langchain-community>=0.2.1,<0.3.0
langchain-anthropic>=0.1.15,<0.2.0
langchain-google-genai>=1.0.5,<2.0.0
langchain-google-firestore>=0.3.0,<0.4.0
firestore>=0.0.8,<0.1.0
chromadb>=0.5.0,<0.6.0
tiktoken>=0.7.0,<0.8.0
sentence-transformers>=3.0.0,<4.0.0
bs4>=0.0.2,<0.1.0
firecrawl-py>=0.0.13,<0.1.0
langchainhub>=0.1.18,<0.2.0
wikipedia>=1.4.0,<2.0.0
tavily-python>=0.3.3,<0.4.0
gradio